micronotif-balance/pyproject.toml
[tool.poetry]
name = "micronotif-balance"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11, <4.0"
redis = {extras = ["hiredis"], version = "^4.5.4"}
grpcio = "^1.54.0"
grpcio-tools = "^1.54.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2-binary = "^2.9.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
